requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
azure-iot-device==2.9.0
azure-iot-hub==2.6.0